Import/Export bot group

+++ lhc_web/design/defaulttheme/js/react/src/components/NodeGroup.js
@@ -46,7 +46,9 @@ class NodeGroup extends Component {
     }
 
     deleteGroup() {
-        this.props.dispatch(deleteGroup({id: this.props.group.get('id')}));
+        if (confirm('Are you sure?')){
+            this.props.dispatch(deleteGroup({id: this.props.group.get('id')}));
+        }
     }
 
     /*shouldComponentUpdate(nextProps, nextState) {
@@ -89,7 +91,7 @@ class NodeGroup extends Component {
                     <hr/>
 
                     <div className="row">
-                        <div className="col-10">
+                        <div className="col-9">
                             <div className="input-group">
                                 <div className="input-group-prepend">
                                     <span><i className={classNameCurrent} title={'Bot Id - '+this.props.group.get('bot_id')}>home</i></span>
@@ -97,8 +99,11 @@ class NodeGroup extends Component {
                                 <input className="form-control form-control-sm gbot-group-name" value={this.props.group.get('name')} onChange={this.handleChange.bind(this)} />
                             </div>
                         </div>
-                        <div className="col-2">
-                            <button className="btn btn-sm btn-secondary float-right" onClick={this.deleteGroup.bind(this)}><i className="material-icons mr-0">delete</i></button>
+                        <div className="col-3">
+                            <div className="btn-group" role="group" aria-label="Basic example">
+                                <a className="btn btn-sm btn-secondary float-right" href={WWW_DIR_JAVASCRIPT + "genericbot/downloadbotgroup/" + this.props.group.get('id')}><i className="material-icons mr-0">cloud_download</i></a>
+                                <button className="btn btn-sm btn-danger float-right" onClick={this.deleteGroup.bind(this)}><i className="material-icons mr-0">delete</i></button>
+                            </div>
                         </div>
                     </div>
